+++ app/src/main/java/com/davidgrath/fitnessapp/YogaTestViewModel.kt
@@ -14,7 +14,7 @@ class YogaTestViewModel(
 //    private val yogaRepository: YogaRepository
 ): ViewModel() {
 
-    var currentWorkoutId: Int = -1
+    var currentWorkoutId: Long = -1
         private set
 
     //TODO this is basically illegal by architecture standards but I'm not abstracting just yet
