+++ app/src/main/java/com/davidgrath/fitnessapp/ui/swimming/SwimmingViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.davidgrath.fitnessapp.data.SwimmingRepository
-import com.davidgrath.fitnessapp.data.entities.SwimmingWorkout
 import com.davidgrath.fitnessapp.data.entities.WorkoutSummary
 import com.davidgrath.fitnessapp.framework.FitnessService
+import com.davidgrath.fitnessapp.framework.database.entities.SwimmingWorkout
 import com.davidgrath.fitnessapp.util.SimpleResult
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.disposables.Disposable
@@ -19,7 +19,7 @@ class SwimmingViewModel(
     private val swimmingRepository: SwimmingRepository
 ) : ViewModel() {
 
-    var currentWorkoutId: Int = -1
+    var currentWorkoutId: Long = -1
     private set
 
     //TODO this is basically illegal by architecture standards but I'm not abstracting just yet
